WSI pietnastka improvements and fixes

+++ sztuczna_inteligencja/pietnastka/heuristics.h
@@ -8,7 +8,7 @@
 #include <cstdint>
 #include "solver.h"
 
-uint naive(BoardEncoded board);
-uint manhattan(BoardEncoded board);
+uint8_t naive(BoardEncoded board);
+uint8_t manhattan(BoardEncoded board);
 
 #endif //PIETNASTKA__HEURISTICS_H_
